@@ -18,27 +18,32 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  google_maps_flutter: ^2.0.0
+  google_maps_flutter: ^2.1.2
   http: ^0.12.2
-  url_launcher: ^5.7.10
+  url_launcher: ^6.0.20
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-  flutter_svg: ^0.21.0+1
+  flutter_svg: ^1.0.3
   shared_preferences: ^2.0.5
   auto_size_text: ^2.1.0
   flutter_linkify: ^5.0.2
   rename: ^1.3.1
   email_validator: ^2.0.1
-  firebase_core: ^1.2.0
-  firebase_messaging: ^10.0.0
   overlay_support: ^1.2.1
+  firebase_messaging: ^11.2.13
+  firebase_core: ^1.14.0
+  cloud_firestore: ^3.1.5
+  notification_permissions: ^0.6.1
+  permission_handler: ^9.2.0
+
+
 
 
 dev_dependencies:
