@@ -21,10 +21,10 @@ class InfoAboutUs extends StatelessWidget {
           end: Alignment.bottomLeft,
           stops: [0.1, 0.5, 0.7, 0.9],
           colors: [
-            Colors.lightGreen[200],
-            Colors.lightGreen[300],
-            Colors.lightGreen[400],
-            Colors.lightGreen[400],
+            Colors.lightGreen[200]!,
+            Colors.lightGreen[300]!,
+            Colors.lightGreen[400]!,
+            Colors.lightGreen[400]!,
           ],
         ),
       ),
